Seperate asset.js, workspace.js, DBpool.js from app.js
